* Made horizontal spacing parameter an "int" instead of a "float" since 	  "float" was making the fonts look blurry. Spacing changes will be 	  choppier, so we will need to revisit that.
git-svn-id: https://iterm.svn.sourceforge.net/svnroot/iterm/trunk@1201 a86d2aa4-6e2d-4ff4-9c10-3c044852e7c9

+++ Headers/iTerm/PseudoTerminal.h
@@ -1,5 +1,5 @@
 // -*- mode:objc -*-
-// $Id: PseudoTerminal.h,v 1.24 2004-05-08 18:26:00 ujwal Exp $
+// $Id: PseudoTerminal.h,v 1.25 2004-05-19 20:38:06 ujwal Exp $
 /*
  **  PseudoTerminal.h
  **
@@ -43,7 +43,7 @@
     
     /////////////////////////////////////////////////////////////////////////
     int WIDTH,HEIGHT;
-	float charWidth; // float for smoother increases
+	int charWidth;
 	int charHeight;
 	float charHorizontalSpacingMultiplier, charVerticalSpacingMultiplier;
     NSFont *FONT, *NAFONT;
